hub-service/requirements.txt
torch
diffusers
transformers
sentencepiece
fastapi
uvicorn
pillow
python-multipart
